1 增加UIMenu控件 2 修复UIGifAnim的gdi+重载错误

+++ Duilib/src/Control/UIGifAnim.cpp
@@ -262,7 +262,7 @@ namespace DuiLib
 		if ( NULL == hDC || NULL == m_pGifImage ) return;
 		GUID pageGuid = Gdiplus::FrameDimensionTime;
 		Gdiplus::Graphics graphics( hDC );
-		graphics.DrawImage( m_pGifImage, m_rcItem.left, m_rcItem.top, m_rcItem.right-m_rcItem.left, m_rcItem.bottom-m_rcItem.top );
+		graphics.DrawImage( m_pGifImage, (int)m_rcItem.left, (int)m_rcItem.top, (int)m_rcItem.right-m_rcItem.left, (int)m_rcItem.bottom-m_rcItem.top );
 		m_pGifImage->SelectActiveFrame( &pageGuid, m_nFramePosition );
 	}
